fix(ai, pty, shell): improve agent tool resiliency and Linux/WSL compatibility
- Coerce stringified numbers and JSON arrays in agent tool schemas (pty_session, read_terminal_buffer, todo_write, code_search)

- Add tool parameter aliases (cmd->command, text/keystrokes->input) and action aliases (exec/execute/shell->run) in repairToolCall

- Fix WSL flag parsing for -e/--exec and key=value args in shell sandbox validation

- Fix REPL pipe reader drain synchronization and until matching on short-lived Linux processes

- Fix Unix PTY process ID type casting and fallback login shell sequence for Linux

- Enforce borderless window state post-initialization on Linux

+++ src-tauri/src/modules/pty/commands.rs
@@ -246,7 +246,7 @@ pub fn pty_has_foreground_job(state: tauri::State<PtyState>, id: u64) -> Result<
     #[cfg(unix)]
     {
         let leader = session.master.lock().unwrap().process_group_leader();
-        Ok(matches!(leader, Some(pid) if pid > 0 && pid as u32 != shell_pid))
+        Ok(matches!(leader, Some(pid) if pid > 0 && (pid as u64) != shell_pid))
     }
     #[cfg(windows)]
     {
